Add loading overlay to login, fix dark mode in Admin/HR/notifications, auto-route users by role, create user from HR form
- Login page: fullscreen three-bar loading overlay during authentication
- AdminPage: add text-obsidian to all <select> elements for dark mode readability
- HRPage: add 'Create Login Account' checkbox + email/password/role fields to Add Employee modal
- DashboardRedirect: auto-route users by role (HR→/hr, Pharmacist→/pharmacy, etc.)
- NotificationBell: increase dropdown to w-96 x max-h-[480px], fix message to 16px text-graphite
- Backend POST /hr/employees: accept createUser flag, atomically create User+Employee in transaction

+++ backend/src/schemas/pos.schema.ts
@@ -6,7 +6,7 @@ export const posTransactSchema = z.object({
   type: z.enum(['PHARMACY', 'OPTICS']),
   items: z.array(z.object({
     id: z.string().min(1),
-    quantity: z.number().int().positive().optional().default(1),
+    quantity: z.number().positive().optional().default(1),
     name: z.string().optional(),
   })).min(1, 'items array is required'),
   paymentMethod: paymentMethods,
